Extract common functionality to DataBranch
